refactor(#391): extract admin users handlers into src/routes/admin_users.rs
admin.rs had drifted to 2042 lines, past the Foundation Rule #12 soft cap
of 2000. Mirror the story 8-4 reference-data extraction: move the entire
Users surface into a new src/routes/admin_users.rs.

Moved (behaviour-preserving):
- handlers: admin_users_panel, admin_users_create_form, admin_users_create,
  admin_users_edit_form, admin_users_deactivate_modal, admin_users_update,
  admin_users_deactivate, admin_users_reactivate
- helpers: validate_username, clamp_users_page, MAX_USERS_PAGE,
  render_users_panel, render_user_row
- query/form/template structs + the corresponding unit tests

The shared admin shell/dispatch (render_admin, render_panel, render_shell,
AdminTab) stays in admin.rs — it serves all five tabs. render_admin is now
pub(crate) and render_users_panel/UsersFilters are pub(crate) so the two
modules can call across the (intentional, single-crate) boundary. Route
registrations in routes/mod.rs updated admin::admin_users_* -> admin_users::*.

admin.rs: 2042 -> 1401 lines. No behaviour change; full unit + integration
suite green (1012 lib tests + admin route integration tests).

Closes #391
